[move] ReferenceRepositoryをcontrollersパッケージに移す
